// apps/web/src/api/generated-api.ts
export interface RegisterBody {
  /** @format email */
  email: string;
  /**
   * @minLength 1
   * @maxLength 64
   */
  username: string;
  /**
   * @minLength 8
   * @maxLength 64
   */
  password: string;
  /**
   * @minLength 8
   * @maxLength 64
   */
  confirmPassword: string;
}